+++ server/core/comms/hub_test.go
@@ -1,14 +1,12 @@
 package comms_test
 
 import (
-	"fmt"
 	"testing"
 	"time"
 
 	"github.com/atulanand206/find/server/core/actions"
 	"github.com/atulanand206/find/server/core/comms"
-	"github.com/atulanand206/find/server/core/models"
-	gonanoid "github.com/matoous/go-nanoid/v2"
+	"github.com/atulanand206/find/server/core/tests"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -20,7 +18,7 @@ func TestHub(t *testing.T) {
 	})
 
 	t.Run("New hub with 1 client", func(t *testing.T) {
-		hub := RunningHub(t)
+		hub := tests.RunningHub(t)
 		client := comms.NewClient(hub, nil)
 		assert.NotNil(t, client, "Client should not be nil")
 		hub.Register <- client
@@ -29,19 +27,19 @@ func TestHub(t *testing.T) {
 	})
 
 	t.Run("New hub with 4 clients", func(t *testing.T) {
-		hub := RunningHub(t)
-		NewClient(t, hub)
-		NewClient(t, hub)
-		NewClient(t, hub)
-		NewClient(t, hub)
+		hub := tests.RunningHub(t)
+		tests.NewClient(t, hub)
+		tests.NewClient(t, hub)
+		tests.NewClient(t, hub)
+		tests.NewClient(t, hub)
 		time.Sleep(10 * time.Millisecond)
 		assert.Equal(t, 4, len(hub.Clients), "Hub should have 4 clients")
 	})
 
 	t.Run("Broadcast Hello to a hub with 4 clients", func(t *testing.T) {
-		hub := RunningHubWithClients(t, 4)
+		hub := tests.RunningHubWithClients(t, 4)
 		assert.NotNil(t, hub, "Hub should not be nil")
-		msg := models.WebsocketMessageCreator{}.InitWebSocketMessage(actions.BEGIN, "Hello")
+		msg := tests.TestMessage(actions.BEGIN, "Hello")
 		targets := make(map[string]bool)
 		for client := range hub.Clients {
 			targets[client.PlayerId] = true
@@ -54,30 +52,4 @@ func TestHub(t *testing.T) {
 			}
 		}
 	})
-}
-
-func RunningHubWithClients(t *testing.T, n int) *comms.Hub {
-	hub := RunningHub(t)
-	for i := 0; i < n; i++ {
-		NewClient(t, hub)
-	}
-	time.Sleep(10 * time.Millisecond)
-	assert.Equal(t, n, len(hub.Clients), fmt.Sprintf("Hub should have %d clients", n))
-	return hub
-}
-
-func RunningHub(t *testing.T) *comms.Hub {
-	hub := comms.NewHub()
-	assert.NotNil(t, hub, "Hub should not be nil")
-	go hub.Run()
-	return hub
-}
-
-func NewClient(t *testing.T, hub *comms.Hub) *comms.Client {
-	client := comms.NewClient(hub, nil)
-	testPlayerId, _ := gonanoid.New(10)
-	client.SetPlayerId(testPlayerId)
-	assert.NotNil(t, client, "Client should not be nil")
-	hub.Register <- client
-	return client
 }
